requirements.txt
psutil
pyip
apscheduler
fastapi-users[sqlalchemy]
fastapi[standard]
uvicorn[standard]
sqlalchemy
databases
pydantic
httpx
torch
torchvision
torchaudio
transformers
accelerate
pydantic_extra_types
pendulum
python-dotenv